chore: migrate api routes to app router
