+++ src/app/api/logger/route.ts
@@ -1,0 +1,1 @@
+export { POST } from "@navikt/next-logger/app-dir";
