--- src/pages/api/logger.ts
@@ -1,1 +1,0 @@
-export { loggingRoute as default } from "@navikt/next-logger/pages";
