+++ vitest.setup.mts
@@ -1,6 +1,6 @@
 import "vitest-dom/extend-expect";
 
-import mockRouter from "next-router-mock";
+import mockRouter, { useRouter } from "next-router-mock";
 import { createDynamicRouteParser } from "next-router-mock/dynamic-routes";
 import pino from "pino";
 import pretty from "pino-pretty";
@@ -90,41 +90,17 @@ mockRouter.useParser((url) => {
 });
 
 vi.mock("graphql", () => vi.importActual("graphql/index.js"));
-vi.mock("next/router", () => vi.importActual("next-router-mock"));
-vi.mock("next/dist/client/router", () => vi.importActual("next-router-mock"));
 
-// Mock next/navigation (App Router) by bridging to next-router-mock (Pages Router mock).
-// The mock reads from the mockRouter singleton at call time, so tests can use
-// mockRouter.setCurrentUrl() to control routing state for all App Router hooks.
-vi.mock("next/navigation", async () => {
-  const { default: router } = await import("next-router-mock");
-  return {
-    useParams: () => router.query,
-    useRouter: () => ({
-      push: (url: string) => router.push(url),
-      replace: (url: string) => router.replace(url),
-      back: () => router.back(),
-      forward: () => {},
-      prefetch: () => Promise.resolve(),
-      refresh: () => {},
-    }),
-    usePathname: () => {
-      const asPath: string = router.asPath;
-      return asPath.split("?")[0];
-    },
-    useSearchParams: () => {
-      const asPath: string = router.asPath;
-      const queryString = asPath.includes("?") ? asPath.split("?")[1] : "";
-      return new URLSearchParams(queryString);
-    },
-    notFound: () => {
-      throw new Error("NEXT_NOT_FOUND");
-    },
-    redirect: (url: string) => {
-      throw new Error(`NEXT_REDIRECT:${url}`);
-    },
-  };
-});
+vi.mock("next/navigation", () => ({
+  useRouter: vi.fn().mockImplementation(useRouter),
+  usePathname: vi.fn().mockImplementation(() => useRouter().pathname),
+  useSearchParams: vi
+    .fn()
+    .mockImplementation(
+      () => new URLSearchParams(useRouter().query as Record<string, string>),
+    ),
+  useParams: vi.fn().mockImplementation(() => useRouter().query),
+}));
 
 // Mock nav-dekoratoren-moduler to prevent timers from running in tests
 vi.mock("@navikt/nav-dekoratoren-moduler", () => ({
